requirements.txt
absl-py==0.9.0
argon2-cffi==20.1.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography @ file:///C:/ci/cryptography_1598873979408/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gast==0.3.3
geographiclib==1.50
geopy==2.0.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.7.1
matplotlib==3.3.1
mistune==0.8.4
mysql-connector-python==8.0.18
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.4
pandas==1.1.0
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.2.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
PyMySQL==0.9.3
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==225
pywinpty==0.5.7
pyxll @ file:///P:/Git/dev_models/pyxll-4.6.0-cp37-none-any.whl
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy @ file:///C:/ci/sqlalchemy_1598374801163/work
statsmodels==0.11.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
Wx @ file:///S:/WX/Models/PROD/wx_wheels/Wx-0.3.0-py3-none-any.whl
zipp==3.1.0
